% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optparse.R
\name{IndentedHelpFormatter}
\alias{IndentedHelpFormatter}
\alias{TitledHelpFormatter}
\title{Builtin help text formatters}
\usage{
IndentedHelpFormatter(object)

TitledHelpFormatter(object)
}
\arguments{
\item{object}{An \code{\link[=OptionParser]{OptionParser()}} object.}
}
\value{
\code{NULL} invisibly.  As a side effect prints out help text.
}
\description{
\code{IndentedHelpFormatter()} is the default help text formatter.
\code{TitledHelpFormatter()} is an alternative help text formatter.
}
\examples{
 parser <- OptionParser(formatter = IndentedHelpFormatter)
 parser <- add_option(parser, "--generator", help = "Generator option")
 parser <- add_option(parser, "--count", help = "Count option")
 print_help(parser)

 parser <- OptionParser(formatter = TitledHelpFormatter)
 parser <- add_option(parser, "--generator", help = "Generator option")
 parser <- add_option(parser, "--count", help = "Count option")
 print_help(parser)
}
