% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{OptionParser-class}
\alias{OptionParser-class}
\title{Option Parser}
\description{
Option Parser
}
\section{Slots}{

\describe{
\item{\code{usage}}{The program usage message that will printed out if
\code{parse_args} finds a help option, \code{\%prog} is substituted with the
value of the \code{prog} argument.}

\item{\code{options}}{A list of of \code{OptionParserOption} instances that will
define how \code{parse_args} reacts to command line options.
\code{OptionParserOption} instances are usually created by \code{make_option}
and can also be added to an existing \code{OptionParser} instance via the
\code{add_option} function.}

\item{\code{description}}{Additional text for \code{print_help} to print out between
usage statement and options statement}

\item{\code{epilogue}}{Additional text for \code{print_help} to print out after
the options statement}
}}
\author{
Trevor Davis.
}
\seealso{
\code{\link{OptionParserOption}}
}

