% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optparse.R
\name{print_help}
\alias{print_help}
\title{Printing an usage message from an OptionParser object}
\usage{
print_help(object)
}
\arguments{
\item{object}{A \code{OptionParser} instance.}
}
\value{
\code{print_help} uses the \code{cat} function to print out a usage
message.  It returns \code{invisible(NULL)}.
}
\description{
\code{print_help} print an usage message from an OptionParser object, usually
called by \code{parse_args} when it encounters a help option.
}
\author{
Trevor Davis.
}
\references{
Python's \code{optparse} library, which inspired this package,
    is described here: \url{http://docs.python.org/library/optparse.html}
}
\seealso{
\code{{parse_args}} \code{{OptionParser}}
}

